+++ src/main/javascript/Context/TicketEventHandlers.js
@@ -7,7 +7,10 @@ import * as Events from './TicketEvents';
 
 /**
  * @method
+ *
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleIncomingEvent( Events.EVENT_TICKET_REPLY, Events.props.EVENT_TICKET_REPLY  )
+export const registerEventHandlers = (windowBridge, app) => {
+  handleIncomingEvent(windowBridge, app, Events.EVENT_TICKET_REPLY, Events.props.EVENT_TICKET_REPLY)
 };