+++ src/main/javascript/Core/EventHandler.js
@@ -3,26 +3,22 @@
  */
 
 import { MessageBus, IncomingEventDispatcher, OutgoingEventDispatcher } from './EventDispatcher';
-import { windowProxy } from './Window';
-import { WidgetRequest, WidgetResponse, parseIncomingMessage, createOutgoingResponseMessage, createOutgoingRequestMessage } from './Message';
-import { on as postRobotOn } from '../../../post-robot';
+import { WidgetRequest, WidgetResponse, WidgetFactories } from '../Widget';
 import { INVOCATION_REQUESTRESPONSE, INVOCATION_FIREANDFORGET } from './Event'
 
+// TODO this should form the basic for the driver for WidgetWindowBridge
+
 /**
+ * @param {WidgetWindowBridge} windowBridge
  * @param {String} eventName
  * @param {Object} eventProps
- * @param rawMessage
+ * @param {{ data:* }} event
  */
-const dispatchIncomingEvent = (eventName, eventProps, rawMessage) => {
-  // this should not be here good here
-  const { onDpMessage: postRobotSend } = windowProxy.xchild.props;
+const dispatchIncomingEvent = (windowBridge, eventName, eventProps, event) => {
+  const { data: rawMessage } = event;
 
-  const message = parseIncomingMessage(rawMessage);
+  const message = WidgetFactories.parseMessageFromJS(rawMessage);
   if (message instanceof WidgetRequest && eventProps.invocationType === INVOCATION_REQUESTRESPONSE) {
-    MessageBus.once(message.correlationId, response => {
-      const payload = response instanceof WidgetResponse ? response.toJS() : response;
-      postRobotSend(eventName, payload);
-    });
     MessageBus.emit(eventName, message);
   }
 
@@ -37,35 +33,31 @@ const dispatchIncomingEvent = (eventName, eventProps, rawMessage) => {
 
 /**
  * TODO: this should handle requests and response, not only requests
- * @param request
+ *
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {String} eventName
+ * @param {WidgetRequest} request
  */
-const createDispatchOutgoingResponseEvent = request => (error, data) => {
-  const responsePayload = error ? error: data;
-  const createErrorResponse = !!error;
-  const response = createOutgoingResponseMessage(request, responsePayload, createErrorResponse);
-
-  MessageBus.emit(response.correlationId, response);
-};
-
-const registerIncomingEventHandler = (eventName, eventProps, eventHandler) => {
-  MessageBus.on(eventName, eventHandler);
-  postRobotOn(eventName, event => dispatchIncomingEvent(eventName, eventProps, event.data));
+const createDispatchOutgoingResponseEvent = (windowBridge, eventName, request) => (error, data) => {
+  windowBridge.emitResponse(eventName, error, data, request).then(({response, emit}) => emit());
 };
 
 /**
  * @method
  *
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {App} app
  * @param {String} eventName
  * @param {Object} eventProps
  */
-export const handleIncomingEvent = (eventName, eventProps) =>
+export const handleIncomingEvent = (windowBridge, app, eventName, eventProps) =>
 {
   let eventHandler;
 
   if (eventProps.invocationType === INVOCATION_REQUESTRESPONSE) {
     eventHandler = request =>
     {
-      const cb = createDispatchOutgoingResponseEvent(request);
+      const cb = createDispatchOutgoingResponseEvent(windowBridge, eventName, request);
       IncomingEventDispatcher.emit(eventName, cb, request.body);
     }
   } else if (eventProps.invocationType === INVOCATION_FIREANDFORGET) {
@@ -73,51 +65,46 @@ export const handleIncomingEvent = (eventName, eventProps) =>
   }
 
   if (eventHandler) {
-    registerIncomingEventHandler(eventName, eventProps, eventHandler);
+    MessageBus.on(eventName, eventHandler);
+    windowBridge.on(eventName, dispatchIncomingEvent.bind(this, windowBridge, eventName, eventProps));
   }
 
-};
-
-// outgoing events
-
-const registerOutgoingEventHandler = (eventName, eventProps, eventHandler) => {
-  OutgoingEventDispatcher.on(eventName, eventHandler);
-  postRobotOn(eventName, event => dispatchIncomingEvent(eventName, eventProps, event.data));
 };
 
 /**
  * @method
  *
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {App} app
  * @param {string} eventName
  * @param {object} eventProps
  */
-export const handleOutgoingEvent = (eventName, eventProps) =>
+export const handleOutgoingEvent = (windowBridge, app, eventName, eventProps) =>
 {
   let eventHandler;
 
   if (eventProps.invocationType === INVOCATION_REQUESTRESPONSE) {
     eventHandler = (resolve, reject, data) => {
-      // here we are sure windowProxy.xchild has been initialized, but this knowledge only makes this function brittle
-      const { widgetId, onDpMessage: postRobotSend } = windowProxy.xchild.props;
-      const request = createOutgoingRequestMessage(widgetId, data);
 
-      // register a response listener
-      MessageBus.once(request.correlationId, response => (response.status == 'success' ? resolve(response.body) : reject(response.body)) );
-      postRobotSend(eventName, request.toJS());
+      windowBridge.emitRequest(eventName, data)
+        .then(({request, emit}) => {
+        // resolve the promise when response comes
+        const responseExecutor = response => (response.status == 'success' ? resolve(response.body) : reject(response.body));
+        MessageBus.once(request.correlationId, responseExecutor);
+        return {request, emit};
+        })
+      .then(({emit}) => emit())
+      ;
     }
   } else if (eventProps.invocationType === INVOCATION_FIREANDFORGET) {
     eventHandler = (resolve, reject, data) => {
-      // here we are sure windowProxy.xchild has been initialized, but this knowledge only makes this function brittle
-      const { widgetId, onDpMessage: postRobotSend } = windowProxy.xchild.props;
-
-      const request = createOutgoingRequestMessage(widgetId, data);
-      postRobotSend(eventName, request.toJS());
-      resolve(data);
+      windowBridge.emitRequest(eventName, data).then(({emit}) => emit()).then(() => resolve(data))
     }
   }
 
   if (eventHandler) {
-    registerOutgoingEventHandler(eventName, eventProps, eventHandler);
+    OutgoingEventDispatcher.on(eventName, eventHandler);
+    windowBridge.on(eventName, dispatchIncomingEvent.bind(this, windowBridge, eventName, eventProps));
   }
 
 };