+++ src/main/javascript/Security/EventHandlers.js
@@ -8,8 +8,11 @@ import { handleOutgoingEvent } from '../Core/EventHandler';
 
 /**
  * @method
+ *
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, Events.props.EVENT_SECURITY_AUTHENTICATE_OAUTH);
-  handleOutgoingEvent(Events.EVENT_SECURITY_SETTINGS_OAUTH, Events.props.EVENT_SECURITY_SETTINGS_OAUTH);
+export const registerEventHandlers = (windowBridge, app) => {
+  handleOutgoingEvent(windowBridge, app, Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, Events.props.EVENT_SECURITY_AUTHENTICATE_OAUTH);
+  handleOutgoingEvent(windowBridge, app, Events.EVENT_SECURITY_SETTINGS_OAUTH, Events.props.EVENT_SECURITY_SETTINGS_OAUTH);
 };