+++ src/main/javascript/WebAPI/EventHandlers.js
@@ -8,8 +8,10 @@ import { handleOutgoingEvent } from '../Core/EventHandler';
 
 /**
  * @method
+ * @param {WidgetWindowBridge} windowBridge
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_WEBAPI_REQUEST_DESKPRO, Events.props.EVENT_WEBAPI_REQUEST_DESKPRO);
-  handleOutgoingEvent(Events.EVENT_WEBAPI_REQUEST_FETCH, Events.props.EVENT_WEBAPI_REQUEST_FETCH);
+export const registerEventHandlers = (windowBridge, app) => {
+  handleOutgoingEvent(windowBridge, app, Events.EVENT_WEBAPI_REQUEST_DESKPRO, Events.props.EVENT_WEBAPI_REQUEST_DESKPRO);
+  handleOutgoingEvent(windowBridge, app, Events.EVENT_WEBAPI_REQUEST_FETCH, Events.props.EVENT_WEBAPI_REQUEST_FETCH);
 };